# use bytes.Buffer instead of string concat