appending to  bytes.Buffer is faster than string concat.

Before this change:
```
$ go test -bench=.
BenchmarkParse-8          500000              3237 ns/op
BenchmarkEval-8          5000000               345 ns/op
PASS
ok      github.com/segmentio/go-interpolate     3.738s
```

After this change:
```
$ go test -bench=.
BenchmarkParse-8         1000000              1743 ns/op
BenchmarkEval-8          5000000               252 ns/op
PASS
ok      github.com/segmentio/go-interpolate     3.290s
```

Parse is typically executed only once so not a big deal. Eval is executed pretty often in our workers so we end up quite a bit here.

Passes locally:
```
$ go test ./...
ok      github.com/segmentio/go-interpolate     0.006s
```