@f2prateek 2017-04-15
Comparison:

```
benchmark            old ns/op     new ns/op     delta
BenchmarkParse-8     3534          1826          -48.33%
BenchmarkEval-8      371           255           -31.27%
```