@f2prateek 2017-04-15
There's another optimization to be made here - which is to call buf.Reset() instead of creating a new one.